Layers

Index

NeuralGraphPDE.AbstractGNNContainerLayer

NeuralGraphPDE.AbstractGNNLayer

NeuralGraphPDE.ExplicitEdgeConv

NeuralGraphPDE.ExplicitGCNConv

NeuralGraphPDE.MPPDEConv

NeuralGraphPDE.VMHConv

Docs

NeuralGraphPDE.AbstractGNNContainerLayer — Type

AbstractGNNContainerLayer <: AbstractExplicitContainerLayer

This is an abstract type of GNN layers that contains other layers.

NeuralGraphPDE.AbstractGNNLayer — Type

AbstractGNNLayer <: AbstractExplicitLayer

An abstract type of graph neural networks. See also AbstractGNNContainerLayer

NeuralGraphPDE.ExplicitEdgeConv — Type

ExplicitEdgeConv(ϕ; initialgraph = initialgraph, aggr = mean)

Edge convolutional layer.

\[\mathbf{h}_i' = \square_{j \in N(i)}\, \phi([\mathbf{h}_i, \mathbf{h}_j; \mathbf{x}_j - \mathbf{x}_i])\]

Arguments

ϕ: A neural network.

initialgraph: GNNGraph or a function that returns a GNNGraph

aggr: Aggregation operator for the incoming messages (e.g. +, *, max, min, and mean).

Inputs

h: Trainable node embeddings, NamedTuple or Array.

Returns

NamedTuple or Array that has the same struct with x with different a size of channels.

Parameters

Parameters of ϕ.

States

graph: GNNGraph where graph.ndata.x represents the spatial coordinates of nodes. You can also put other nontrainable node features in graph.ndata with arbitrary keys. They will be concatenated like u.

Examples

s = [1, 1, 2, 3]
t = [2, 3, 1, 1]
g = GNNGraph(s, t)

u = randn(4, g.num_nodes)
g = GNNGraph(g, ndata = (; x = rand(3, g.num_nodes)))
nn = Dense(4 + 4 + 3 => 5)
l = ExplicitEdgeConv(nn, initialgraph=g)

rng = Random.default_rng()
ps, st = Lux.setup(rng, l)

NeuralGraphPDE.ExplicitGCNConv — Type

ExplicitGCNConv(in_chs::Int, out_chs::Int, activation = identity;
                initialgraph = initialgraph, init_weight = glorot_normal,
                init_bias = zeros32)

Same as the one in GraphNeuralNetworks.jl but with exiplicit paramters.

Arguments

initialgraph: GNNGraph or a function that returns a GNNGraph

Examples

# create data
s = [1,1,2,3]
t = [2,3,1,1]
g = GNNGraph(s, t)
x = randn(3, g.num_nodes)

# create layer
l = ExplicitGCNConv(3 => 5, initialgraph = g) 

# setup layer
rng = Random.default_rng()
Random.seed!(rng, 0)

ps, st = Lux.setup(rng, l)

# forward pass
y = l(x, ps, st)       # size:  5 × num_nodes

NeuralGraphPDE.MPPDEConv — Type

MPPDEConv(ϕ, ψ; initialgraph = initialgraph, aggr = sum, local_features = (:u, :x))

Convolutional layer from Message Passing Neural PDE Solvers.

\[\begin{aligned} \mathbf{m}_i&=\Box _{j\in N(i)}\,\phi (\mathbf{h}_i,\mathbf{h}_j;\mathbf{u}_i-\mathbf{u}_j;\mathbf{x}_i-\mathbf{x}_j;\theta )\\ \mathbf{h}_i'&=\psi (\mathbf{h}_i,\mathbf{m}_i,\theta )\\ \end{aligned}\]

Arguments

ϕ: The neural network for the message function.

ψ: The neural network for the update function.

initialgraph: GNNGraph or a function that returns a GNNGraph

aggr: Aggregation operator for the incoming messages (e.g. +, *, max, min, and mean).

local_features: The features that will be differentiated in the message function.

Inputs

h: Trainable node embeddings, Array.

Returns

NamedTuple or Array that has the same struct with x with different a size of channels.

Parameters

Parameters of ϕ.

Parameters of ψ.

States

graph: GNNGraph where graph.ndata.x represents the spatial coordinates of nodes, and graph.gdata.θ represents the graph level features. θ should be a vector.

Examples

g = rand_graph(10, 6)

g = GNNGraph(g, ndata = (; u = rand(2, 10), x = rand(3, 10)), gdata = (; θ = rand(4)))
h = randn(5, 10)
ϕ = Dense(5 + 5 + 2 + 3 + 4 => 5)
ψ = Dense(5 + 5 + 4 => 7)
l = MPPDEConv(ϕ, ψ, initialgraph = g)

rng = Random.default_rng()
ps, st = Lux.setup(rng, l)

y, st = l(h, ps, st)

NeuralGraphPDE.VMHConv — Type

VMHConv(ϕ, γ; initialgraph = initialgraph, aggr = mean)

Convolutional layer from Learning continuous-time PDEs from sparse data with graph neural networks.

\[\begin{aligned} \mathbf{m}_i &= \square_{j \in N(i)}\, \phi(\mathbf{h}_i, \mathbf{h}_j - \mathbf{h}_i; \mathbf{x}_j - \mathbf{x}_i)\\ \mathbf{h}_i' &= \gamma(\mathbf{h}_i ,\mathbf{m}_i) \end{aligned}\]

Arguments

ϕ: The neural network for the message function.

γ: The neural network for the update function.

initialgraph: GNNGraph or a function that returns a GNNGraph

aggr: Aggregation operator for the incoming messages (e.g. +, *, max, min, and mean).

Inputs

h: Trainable node embeddings, NamedTuple or Array.

Returns

NamedTuple or Array that has the same struct with x with different a size of channels.

Parameters

Parameters of ϕ.

Parameters of γ.

States

graph: GNNGraph where graph.ndata.x represents the spatial coordinates of nodes.

Examples

s = [1, 1, 2, 3]
t = [2, 3, 1, 1]
g = GNNGraph(s, t)

u = randn(4, g.num_nodes)
g = GNNGraph(g, ndata = (; x = rand(3, g.num_nodes)))
ϕ = Dense(4 + 4 + 3 => 5)
γ = Dense(5 + 4 => 7)
l = VMHConv(ϕ, γ, initialgraph = g)

rng = Random.default_rng()
ps, st = Lux.setup(rng, l)

y, st = l(u, ps, st)